requirements.txt
Django
django-annoying
django-classy-tags
django-cms
djangocms-attributes-field
djangocms-file
djangocms-link
djangocms-picture
djangocms-text-ckeditor
django-contact-form
django-filer
django-formtools
django-haystack
django-localflavor
django-mptt
django-nocaptcha-recaptcha
django-parler
django-polymorphic
django-pure-pagination
django-registration
django-sekizai
easy-thumbnails
emoji
GDAL
html5lib
ipaddress
Markdown
mock
mysql-connector-python
Pillow
pytils
simplejson
six
sorl-thumbnail
